refactor: remove legacy code

+++ src/zetton_inference/vision/base/transform/pad.cc
@@ -17,7 +17,7 @@ bool Pad::RunOnOpenCV(Mat* mat) {
     AERROR_F("Pad: Only support channels <= 4.");
     return false;
   }
-  if (mat->Channels() != value_.size()) {
+  if (mat->Channels() != static_cast<int>(value_.size())) {
     AERROR_F(
         "Pad: Require input channels equals to size of padding value, "
         "but now channels = {}, the size of padding values = {}.",
